requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cloudinary==1.44.2
cryptography==47.0.0
dj-rest-auth==7.2.0
Django==6.0.4
django-allauth==65.16.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
idna==3.13
inflection==0.5.1
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.2
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.2
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
